+++ src/lib/bundled-plugins/generic-importer.js
@@ -1,31 +1,29 @@
-import {resolve as resolvePath, isAbsolute} from 'path';
 import {readFile} from 'sander';
-
+import path from 'path';
 
 export default function (baseDir) {
-  function exhibitGenericImporter(path) {
+  function exhibitGenericImporter(file) {
     const accessed = new Set();
     const result = {accessed};
 
     if (!baseDir) {
-      if (!isAbsolute(path)) return result; // can't handle non-absolute paths without a base dir set
+      if (!path.isAbsolute(file)) return result; // can't handle non-absolute paths without a base dir set
     }
-    else path = resolvePath(baseDir, path);
+    else file = path.resolve(baseDir, file);
 
-    console.assert(isAbsolute(path), 'should be absolute now');
+    console.assert(path.isAbsolute(file), 'should be absolute now');
 
-    accessed.add(path);
-    return readFile(path).then(contents => {
+    accessed.add(file);
+
+    return readFile(file).then(contents => {
       result.contents = contents;
-      result.path = path;
+      result.file = file;
       return result;
     }).catch(error => {
       if (error.code !== 'ENOENT' && error.code !== 'EISDIR') throw error;
-
       return result;
     });
   }
-
 
   if (baseDir) {
     Object.defineProperty(exhibitGenericImporter, 'dir', {value: baseDir});