+++ src/lib/cache.js
@@ -1,4 +1,4 @@
-import {basename, resolve as resolvePath, normalize, dirname, relative, isAbsolute} from 'path';
+import path from 'path';
 import {lstat, readFile, readdir, writeFile, unlink} from 'sander';
 import {VirtualFolder} from 'exhibit-core';
 import mkdirp from 'mkdirp-then';
@@ -11,26 +11,26 @@ const EXISTING_DIR_PROMISES = Symbol();
 
 /**
  * Traverses and loads all contents of all files under `dir`.
- * TODO: make safer, eg. cap how many files/bytes etc, probably controlled by options.
  */
 const recursiveLoadDir = async dir => {
+  // TODO: make this safer, eg. cap how many files/bytes etc, probably controlled by .build() options.
   const finalResults = {};
 
-  await Promise.map(readdir(dir), async path => {
-    if (basename(path) === '.DS_Store') return;
+  await Promise.map(readdir(dir), async file => {
+    if (path.basename(file) === '.DS_Store') return;
 
-    path = resolvePath(dir, path);
+    file = path.resolve(dir, file);
 
-    const stat = await lstat(path);
+    const stat = await lstat(file);
 
     if (stat.isFile()) {
-      finalResults[path] = await readFile(path);
+      finalResults[file] = await readFile(file);
     }
     else if (stat.isDirectory()) {
-      const results = await recursiveLoadDir(path);
-      Object.keys(results).forEach(path2 => finalResults[path2] = results[path2]);
+      const results = await recursiveLoadDir(file);
+      Object.keys(results).forEach(file2 => finalResults[file2] = results[file2]);
     }
-    else throw new Error('Not sure what to do for file ' + path);
+    else throw new Error('Not sure what to do for file ' + file);
   });
 
   return finalResults;
@@ -47,7 +47,9 @@ export default class Cache extends VirtualFolder {
     };
   }
 
-
+  /**
+   * Instruct the cache to prime itself by loading in all contents from the disk directory.
+   */
   async prime() {
     const cache = this;
 
@@ -58,28 +60,28 @@ export default class Cache extends VirtualFolder {
 
     return Promise.map(
       Object.keys(results),
-      path => {
-        console.assert(isAbsolute(path), 'Expected recursiveLoadDir() to return only absolute paths; got: ' + path);
+      file => {
+        console.assert(path.isAbsolute(file), 'Expected recursiveLoadDir() to return only absolute paths; got: ' + file);
 
-        return super.write(relative(this[DISK_DIR], path), results[path]);
+        return super.write(path.relative(this[DISK_DIR], file), results[file]);
       }
     );
   }
 
 
   /**
-   * Sets the contents for `path` and persists any resulting change
+   * Sets the contents for `file` and persists any resulting change
    * to disk.
    */
-  async write(path, contents) {
-    console.assert(!isAbsolute(path), 'Do not write absolute paths to Cache instances.');
+  async write(file, contents) {
+    console.assert(!path.isAbsolute(file), 'Do not write absolute paths to Cache instances.');
 
-    this.emit('writing', path);
+    this.emit('writing', file);
 
-    const change = super.write(path, contents);
+    const change = super.write(file, contents);
 
     if (change) {
-      const absolutePath = resolvePath(this[DISK_DIR], change.path);
+      const absolutePath = path.resolve(this[DISK_DIR], change.file);
 
 
       if (change.type === 'delete') {
@@ -87,7 +89,7 @@ export default class Cache extends VirtualFolder {
         // todo: delete directory too if it's now empty
       }
       else {
-        await this.ensureDirExists(dirname(change.path));
+        await this.ensureDirExists(path.dirname(change.file));
         await writeFile(absolutePath, change.contents);
       }
     }
@@ -100,19 +102,19 @@ export default class Cache extends VirtualFolder {
    * Allows setting the in-memory virtual folder contents without persisting
    * to disk.
    */
-  async writeWithoutPersisting(path, contents) {
-    console.assert(!isAbsolute(path), 'Do not write absolute paths to Cache instances.');
+  async writeWithoutPersisting(file, contents) {
+    console.assert(!path.isAbsolute(file), 'Do not write absolute paths to Cache instances.');
 
-    return super.write(path, contents);
+    return super.write(file, contents);
   }
 
 
   ensureDirExists(dir) {
     const destination = this;
-    dir = normalize(dir);
+    dir = path.normalize(dir);
 
     if (!destination[EXISTING_DIR_PROMISES][dir]) {
-      destination[EXISTING_DIR_PROMISES][dir] = mkdirp(resolvePath(this[DISK_DIR], dir));
+      destination[EXISTING_DIR_PROMISES][dir] = mkdirp(path.resolve(this[DISK_DIR], dir));
     }
 
     return destination[EXISTING_DIR_PROMISES][dir];