+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2015-10-12
+### Added
+- Generator support: builders can now be provided as generator functions, in which case they are automatically wrapped as [Bluebird coroutines](https://github.com/petkaantonov/bluebird/blob/master/API.md#generators).
+
+### Changed
+- New API for builders: no more `this`; and builders now take a single argument – a 'job' containing data about what needs building. And file paths now have the property name `file`, not `path` (to avoid clashing with Node's path module).
+- Better readme.
+
+
+## [0.5.0] - 2015-10-06
+### Changed
+- Builder plugins are now `exhibit-builder-*`
+- Fledgling docs
+- Restructured: no more multiform; now requiring Node 4+
+
 
 ## [0.4.0] - 2015-07-08
 ### Changed
@@ -27,5 +42,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 [unreleased]: https://github.com/exhibitjs/exhibit/compare/v0.4.0...HEAD
+[0.6.0]: https://github.com/exhibitjs/exhibit/compare/v0.5.0...v0.6.0
+[0.5.0]: https://github.com/exhibitjs/exhibit/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/exhibitjs/exhibit/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/exhibitjs/exhibit/compare/v0.1.0...v0.3.0